+++ packages/units/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @baustatik/units
 
+## 0.1.0
+
+### Minor Changes
+
+- refactor package. move round functions into own package
+
+### Patch Changes
+
+- Updated dependencies
+  - @baustatik/round@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes